requirements.txt
python-dateutil>=2.9.0
joblib>=1.4.0
matplotlib>=3.8.0
numpy>=1.26.0
pandas>=2.2.0
pydantic>=2.6.0
pyyaml>=6.0.0
pytest>=8.0.0
scikit-learn>=1.4.0
scipy>=1.12.0
